Compile thunk actions into organized files
